Relocate redis config
